Optimize Layout (and Drawing) Of Widgets
Previously GLPK (through glpsol) was used to create a linear program
to solve for the layout using a series of global constraints.
This approach had quite a few nice features in terms of overall
constraint flexibility, though it was slow and relatively hard to
debug.
Previous versions could easily take 150+ ms (10+ missed frames) to
perform layout, using a more direct approach results in ~1ms for
each layout.
There are some layout changes in this and there is a new optimized
set of drawing routines in src/quickdraw.c to help with lag induced
when animation frames resulted in changes in plots.