Knob.label is not ignored anymore.
In the case where a knob is attached to a remote value, the short name
of that remote value is set as the label, overwriting whichever value
was set when constructing the object.

Visible impact prior to this change: in the macro learn panel, the
offset knobs were labeled "off" despite having a label in
ZynAutomationParam.qml set to "offset".